+++ Providers/EFCore/Revo.EFCore/DataAccess/Entities/RequestDbContextCache.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using Microsoft.EntityFrameworkCore;
 
 namespace Revo.EFCore.DataAccess.Entities